+++ src/cmd/go/internal/modfetch/fetch.go
@@ -25,10 +25,10 @@ import (
 	"cmd/go/internal/fsys"
 	"cmd/go/internal/gover"
 	"cmd/go/internal/lockedfile"
-	"cmd/go/internal/par"
 	"cmd/go/internal/robustio"
 	"cmd/go/internal/str"
 	"cmd/go/internal/trace"
+	"cmd/internal/par"
 
 	"golang.org/x/mod/module"
 	"golang.org/x/mod/sumdb/dirhash"
