+++ src/cmd/go/internal/work/shell.go
@@ -10,8 +10,8 @@ import (
 	"cmd/go/internal/cache"
 	"cmd/go/internal/cfg"
 	"cmd/go/internal/load"
-	"cmd/go/internal/par"
 	"cmd/go/internal/str"
+	"cmd/internal/par"
 	"errors"
 	"fmt"
 	"internal/lazyregexp"
