+++ src/cmd/internal/par/queue.go
