+++ src/cmd/internal/par/work.go
