+++ src/cmd/internal/par/work_test.go
